feat(calculator) : 커스텀 구분자를 분리해내는 기능
입력된 값에서 커스텀 구분자 입력값일 시, 커스텀 구분자를 추출하는 기능 추가
