+++ src/main/java/calculator/Application.java
@@ -6,7 +6,7 @@ public class Application {
     public static void main(String[] args) {
 
         System.out.println("덧셈할 문자열을 입력해 주세요.");
-        Carculator.sumNumbersFromLetters(Console.readLine());
+        Calculator.sumNumbersFromLetters(Console.readLine());
 
     }
 }
